requirements.txt
torch==2.2.0
torchvision==0.17.0
torchaudio==2.2.0
soundfile
numpy==1.26.4
torch_audiomentations
tqdm
matplotlib
pandas
wandb
comet_ml
hydra-core

black
isort
pre-commit
wget
itertools
transfromers
pytorch_lightning